requirements.txt
asgiref==3.3.0
certifi==2020.6.20
Django==3.1.3
django-allow-cidr==0.3.1
django-js-asset==1.2.2
django-mptt==0.11.0
netaddr==0.8.0
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
whitenoise==5.2.0
